# server/config/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Mean Application API",
    "description": "Mean Application API",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:4040",
  "basePath": "/api/",
  "tags": [
    {
      "name": "Users",
      "description": "API for users in the system"
    },
    {
      "name": "Auth",
      "description": "API for auth in the system"
    }
  ],
  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "schema": "bearer"
    }
  },  
  "security": [
    {
      "Bearer": []
    }
  ],
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["Auth"],
        "description": "Resgister user",
        "parameters": [
          {
            "name": "auth",
            "in": "body",
            "description": "User auth details",
            "schema": {
              "type": "object",
              "required": ["firstName", "lastName", "middleName", "email", "password"],
              "properties": {
                "firstName": {
                  "type": "string"
                },
                "lastName": {
                  "type": "string"
                },
                "middleName": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "User is loggedin",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "description": "Login to the system",
        "parameters": [
          {
            "name": "auth",
            "in": "body",
            "description": "User auth details",
            "schema": {
              "type": "object",
              "required": ["email", "password"],
              "properties": {
                "email": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "User is loggedin",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": ["Auth"],
        "description": "get user info",
        "parameters": [
          {
            "name": "auth",
            "in": "body",
            "description": "User auth details"
          } 
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "User",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/user/contacts/add": {
      "post": {
        "tags": ["User"],
        "description": "get user info",
        "parameters": [
          {
            "name": "add contact",
            "in": "body",
            "schema": {
              "type": "object",
              "required": ["email", "firstName", "lastName", "mobile"],
              "properties": {
                "userId": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "firstName": {
                  "type": "string"
                },
                "lastName": {
                  "type": "string"
                },
                "mobile": {
                  "type": "string"
                }
              }
            }
          } 
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "User",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/user/contacts": {
      "get": {
        "tags": ["User"],
        "description": "get user info",
        "parameters": [
          {
            "name": "get contacts"
          } 
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "User",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/user/contact/{id}": {
      "delete": {
        "tags": ["User"],
        "description": "get user info",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
             "type": "string"
            }
          } 
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "User",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "User": {
      "required": ["email", "firstName", "lastName"],
      "properties": {
        "_id": {
          "type": "string",
          "uniqueItems": true
        },
        "email": {
          "type": "string",
          "uniqueItems": true
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        },
        "roles": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "Users": {
      "type": "array",
      "$ref": "#/definitions/User"
    },
    "Auth": {
      "type": "object",
      "properties": [
        {
          "token": {
            "type": "string"
          },
          "user": {
            "$ref": "#/definitions/User"
          }
        }
      ]
    }
  }
}